authetication refactored
